+++ src/us/kbase/workspace/kbase/KBaseWorkspaceConfig.java
@@ -22,7 +22,7 @@ import software.amazon.awssdk.regions.Region;
 
 public class KBaseWorkspaceConfig {
 	
-	//TODO CODE How the config is created could use a rethink. Would be much simpler just throwing an exception rather than collecting errors.
+	//TODO CONFIG How the config is created could use a rethink. Would be much simpler just throwing an exception rather than collecting errors.
 	/* I think it was originally this way because Glassfish would do (?? I forget) when
 	 * the WS threw an exception, which we wanted to avoid, and so we made the service start
 	 * but throw errors for all calls (e.g. "the service did not start correctly, check the logs".
@@ -30,7 +30,13 @@ public class KBaseWorkspaceConfig {
 	 * On the other hand, it's nice to see all the errors at once *shrug*
 	 */
 	
+	// TODO CONFIG the config docs include the port, but the port is ignored in deploy.cfg
+	//    The tomcat cutover didn't handle the documentation update apparently
+	//    I would prefer if all the configuration was in one place, namely deploy.cfg,
+	//    rather than scattered around multiple files
 	//TODO JAVADOCS
+	//TODO CODE use optionals instead of nulls.
+	//TODO CODE consider returning classes containing related parameters rather than individual parameters.
 	
 	//required deploy parameters
 	private static final String HOST = "mongodb-host";
@@ -60,10 +66,15 @@ public class KBaseWorkspaceConfig {
 	private static final String KBASE_AUTH_ADMIN_FULL_ROLES =
 			"auth2-ws-admin-full-roles";
 	
-	// shock info 
+	// shock / blobstore info 
 	private static final String BYTESTREAM_USER = "bytestream-user";
 	private static final String BYTESTREAM_TOKEN = "bytestream-token";
 	private static final String BYTESTREAM_URL = "bytestream-url";
+	
+	// sample service info 
+	private static final String SAMPLE_SERVICE_URL = "sample-service-url";
+	private static final String SAMPLE_SERVICE_TOKEN = "sample-service-administrator-token";
+	private static final String SAMPLE_SERVICE_TAG = "sample-service-tag";
 	
 	//handle service info
 	private static final String IGNORE_HANDLE_SERVICE = "ignore-handle-service";
@@ -107,6 +118,9 @@ public class KBaseWorkspaceConfig {
 	private final URL bytestreamURL;
 	private final String bytestreamUser;
 	private final String bytestreamToken;
+	private final URL sampleServiceURL;
+	private final String sampleServiceToken;
+	private final String sampleServiceTag;
 	private final String workspaceAdmin;
 	private final String mongoUser;
 	private final String mongoPassword;
@@ -251,6 +265,20 @@ public class KBaseWorkspaceConfig {
 						BYTESTREAM_USER, BYTESTREAM_TOKEN, BYTESTREAM_URL));
 			}
 		}
+		
+		sampleServiceURL = getUrl(config, SAMPLE_SERVICE_URL, paramErrors, false);
+		if (sampleServiceURL == null) {
+			sampleServiceToken = null;
+			sampleServiceTag = null;
+		} else {
+			sampleServiceToken = nullIfEmpty(config.get(SAMPLE_SERVICE_TOKEN));
+			if (sampleServiceToken == null) {
+				paramErrors.add(String.format("If %s is supplied, %s is required",
+						SAMPLE_SERVICE_URL, SAMPLE_SERVICE_TOKEN));
+			}
+			sampleServiceTag = nullIfEmpty(config.get(SAMPLE_SERVICE_TAG));
+		}
+		
 		workspaceAdmin = nullIfEmpty(config.get(WSADMIN));
 		
 		final String muser = nullIfEmpty(config.get(MONGO_USER));
@@ -384,6 +412,9 @@ public class KBaseWorkspaceConfig {
 		if (bytestreamURL != null) {
 			paramSet.addAll(Arrays.asList(BYTESTREAM_URL, BYTESTREAM_USER));
 		}
+		if (sampleServiceURL != null) {
+			paramSet.addAll(Arrays.asList(SAMPLE_SERVICE_URL, SAMPLE_SERVICE_TAG));
+		}
 		for (final String s: paramSet) {
 			if (!nullOrEmpty(cfg.get(s))) {
 				// TODO CODE this should probably be the actual instance value so defaults
@@ -504,6 +535,18 @@ public class KBaseWorkspaceConfig {
 	
 	public String getBytestreamToken() {
 		return bytestreamToken;
+	}
+	
+	public URL getSampleServiceURL() {
+		return sampleServiceURL;
+	}
+	
+	public String getSampleServiceToken() {
+		return sampleServiceToken;
+	}
+	
+	public String getSampleServiceTag() {
+		return sampleServiceTag;
 	}
 	
 	public String getWorkspaceAdmin() {